requirements.txt
absl-py==1.4.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array-record==0.4.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.3.1
certifi==2024.7.4
cffi==1.17.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
command-not-found==0.3
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.11
dbus-python==1.2.16
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distro==1.4.0
distro-info==0.23+ubuntu1.1
dm-tree==0.1.8
etils==1.3.0
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
flatbuffers==23.5.26
fonttools==4.53.1
fqdn==1.5.1
gast==0.4.0
gh==0.0.4
gin-config==0.5.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.21.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.56.0
h11==0.14.0
h5py==3.9.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==2.8
imageio==2.35.0
immutabledict==4.2.0
importlib-metadata==6.7.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaggle==1.6.17
keras==2.13.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==16.0.0
lxml==5.3.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.5
matplotlib-inline==0.1.7
mediapy==1.2.2
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
overrides==7.7.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
portalocker==2.10.1
prometheus_client==0.20.0
promise==2.3
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.22
Pygments==2.18.0
PyGObject==3.36.0
pyparsing==3.1.2
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==26.1.0
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.7.24
remotezip==0.12.3
requests==2.32.3
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
rsa==4.9
sacrebleu==2.4.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.3
sentencepiece==0.2.0
seqeval==1.2.2
six==1.14.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
ssh-import-id==5.10
stack-data==0.6.3
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.1
tensorflow-datasets==4.9.2
tensorflow-estimator==2.13.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.14.0
tensorflow-model-optimization==0.8.0
tensorflow-text==2.13.0
termcolor==2.3.0
terminado==0.18.1
text-unidecode==1.3
tf-keras==2.15.0
tf-models-official==2.13.2
tf-slim==1.1.0
threadpoolctl==3.5.0
tifffile==2023.7.10
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.5.0
tzdata==2024.1
ubuntu-pro-client==8001
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.25.8
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.3.6
widgetsnbextension==4.0.13
wrapt==1.15.0
zipp==3.15.0
